# app/src/swagger/swagger_spec.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.1",
        "title": "MethodFitness"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/auth": {
            "post": {
                "x-name": "loggin",
                "description": "checks credentials",
                "operationId": "auth",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/signIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/auth"
                        }
                    }
                }
            }
        },
...
        "/trainer/hireTrainer": {
            "post": {
                "x-name": "/trainer/hireTrainer",
                "description": "hire trainer",
                "operationId": "/trainer/hireTrainer",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trainer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    }
                }
            }
        },
        "/trainer/updateTrainerInfo": {
            "post": {
                "x-name": "/trainer/updateTrainerInfo",
                "description": "update Trainer Info",
                "operationId": "/trainer/updateTrainerInfo",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateTrainerInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/trainer/updateTrainerContact": {
            "post": {
                "x-name": "/trainer/updateTrainerContact",
                "description": "update Trainer Contact",
                "operationId": "/trainer/updateTrainerContact",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateTrainerContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/trainer/updateTrainerAddress": {
            "post": {
                "x-name": "/trainer/updateTrainerAddress",
                "description": "update Trainer Address",
                "operationId": "/trainer/updateTrainerAddress",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateTrainerAddress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/trainer/updateTrainerPassword": {
            "post": {
                "x-name": "/trainer/updateTrainerPassword",
                "description": "update Trainer Password",
                "operationId": "/trainer/updateTrainerPassword",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateTrainerPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/trainer/updateTrainersClients": {
            "post": {
                "x-name": "/trainer/updateTrainersClients",
                "description": "update Trainers Clients",
                "operationId": "/trainer/updateTrainersClients",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateTrainersClients"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/trainer/getTrainer/{id}": {
            "get": {
                "x-name": "trainer",
                "description": "retrieve single trainer by id",
                "operationId": "trainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The trainers id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
...
        "/client/addClient": {
            "post": {
                "x-name": "/client/addClient",
                "description": "add client",
                "operationId": "/client/addClient",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/updateClientInfo": {
            "post": {
                "x-name": "/client/updateClientInfo",
                "description": "update Client Info",
                "operationId": "/client/updateClientInfo",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateClientInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "signIn": {
            "title": "signIn",
            "properties": {
                "userName": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "x-input": "password"
                }
            },
            "required": [
                "userName",
                "password"
            ]
        },
...
        "auth": {
            "title": "auth",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "user": {
                    "type": "object",
                    "title": "authUser",
                    "properties": {
                        "id": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "role": {
                            "type": "string"
                        },
                        "active": {
                            "type": "boolean"
                        },
                        "userName": {
                            "type": "string"
                        }
                    }
                }
            },
            "required": [
                "success",
                "user"
            ]
        },
...
        "standardSuccessResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "payload": {
                    "type": "object"
                }
            }
        },
...
    },
...
    "securityDefinitions": {},
...
}